'gosecure' is correctly parsed
For the moment, the 'gosecure' keyword behaves as the 'go' keyword.

Next step is to compute transitive closure of the expression on the right and emit dedicated assembly.
For the moment, we will allow all calls/accesses from the function and clone the transitive closure into a separate portion of the assembly generated.