Trying to detect cross domain calls to rcv/send
Added isencl to channels and routines.
When performing a send or a recv, we can find cases when boundaries
between the two domains are crossed.

The next step is to enqueue a fake sudog in the corresponding domain.
This will be tricky as crossdomain allocation is not allowed.
We will need to have a pool of such objects.