New sgqueue implementation
I had to give up on the lock free structure for the moment.
The problem is that with multiple consummers/producers, we run into an
issue where an extra array is required to sync operations (much like the
P structure).
I will try to explore this design latter but decided to start with a
simple spinlock and woundAndDie approach. Non critical operations are
best effort and may fail.
The next approach for performance would be to add a per P queue and to
dispatch uniformly the sudogs to them.

lock free structures are the last step for optimization.