Saving
Implemented write barrier compatible queues.
Does seems to reduce the live-lock problem but does not remove it.
Next: measure the schedticks per second when we do not run the
avoidDeadlock routine. Then try a per P queue with a P that cannot block
if some routine crossed the boundary.
Other issue might be that the enclave itself is running low on processes
and increases the delay to migrate routines.
One observation that we made is that the queue seems to be empty most of
the time.