Performance is shit :bug:
The performance went down, apparently the latency is way up.
I need to look into that. Another option is to start back from the
measurements and then apply selective changes.

Finally, last option is to have both queue implementations and compare
them. It seems that the current ones have an overhead (latency went up
to 4.3us!)