:ambulance: :bug: Fixing futex for good.
The issue of live-lock experienced was that all the threads could go on a
futex_sleep due to the enclave holding a lock (more often than not, a
lock to a channel).
The solution was to try and prevent this behaviour by differentiating
whether or not the lock was held by the enclave (see lock_futex).
It might still be the case that we f*cked up between the three possible
values taken by the lock (sleep, locked, locked_enclave) and might have
to come back to that later on.
The g structure now has a markednofutex attribute to prevent it from
sleeping.
A better option would be to mark channels that are used for crossdomain
communications and prevent anyone from doing a sleep on their locks.
EXPLORE that :fire:.

Then, the second issue was that preemption could happen while holding a
secspinlock, leaving others spinning like cray-cray forever. To prevent
this, I simply increase/decrease the g.m.locks in lock/unlock.

So far, it seems to fix the problem (> 25 runs of live-lock code and >
10 runs of throughput without seeing the issue).