:bookmark: Still the same problem
1. I measured the cache misses, it seems to be fine (actually better
than regular go).
2. Measurements seems to show that the global queue is not drained often
enough in the non-secured env. That seems to be because the scheduler is
not executed often enough.
3. The contention on the queues seems to be fine, we have failing rates
of < 2% most of the time.

So what can we do about it?
    * Maybe new lock-less producer/consumer queues?
    * Maybe a per-p queue and counter? (see below)

Take the problem differently, look at the compilation phase.
Mark the channels used between the two domains somehow.
If you're gonna block on a channel between the two, increment a per-p
counter, keep the p spinning while counter is not 0. This p is just
going to poll the queue. This actually does not work if p blocks on a
futex or syscall later on (kind of bad even!).

Other option, dedicate a p spinning that polls the queue. Better, but
might create contention on the queues.

Last option, create better queues? Put the unlocked routine closer to
the corresponding scheduler maybe. I need to study how the throughput
code behaves, i.e., if there is someone spinning at some point.