:droplet: Cleaning a little before multithreading
For multithreading, I need to call clone with nil nil for m and g, and
fn must be an assembly function that does a thing similar to what was
performed by the original sgx code, just a little simplified.
The challenge now is to find the mstart address inside the enclave.