:bookmark: Fixing M tracing in enclave + stack
Problem 1: stack shrinking relocates pointers and needs some patching.
See the diff in stack.go.
The bug that we observed was that an enclave routine was reading a nil
value from the channel (due to relocation of the recepient).

Problem 2: The enclave M was not part of the mark root set (since it was
allocated as an extra page for the enclave). As a result, we were
loosing objects during collection (e.g., m.gsignal). To fix it, we have
a stub array that is part of the bss and is supposed to have the proper
size of number_tcs * (msgx_size + tls_size) + padding for alignment.
The gosec package was modified to use this bss area as the TLS for
enclave threads. Seems to work but still need to make sure stuff
overthere gets scanned.

The simulation does not seem to error anymore, but the enclave still
does.